Remove stringToIntegerMD5 from ValidTransforms
Summary:
stringToIntegerMD5 was originally intended to hash user agents for privacy, and
possibly strings to use as distkeys.  It is unused and the ingester doesn't
know it exists, so it's safer to destroy it.

Test Plan:
Unit tests pass; will run integration tests on Blueprint and Spade after
upgrading them.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5420